+++ src/main/java/com/example/javagroupproject1/repository/IRepository.java
@@ -2,7 +2,7 @@ package com.example.javagroupproject1.repository;
 
 public interface IRepository<T> {
     Iterable<T> GetAll();
-    T GetByID(int id);
+    T GetByID(long id);
     void Add(T entity);
     void Update(T entity);
     void Delete(T entity);
